working on metabolism. separate metabolism plots
